Issue #178 add new notification center service
