+++ Entangle/entangle-symfony/public/src/entangle/app/config/config.yml
@@ -68,3 +68,11 @@ swiftmailer:
     username:  "%mailer_user%"
     password:  "%mailer_password%"
     spool:     { type: memory }
+
+
+services:
+  notification_center.service:
+   class: Megasoft\EntangleBundle\Classes\NotificationCenter
+   arguments:
+        em: "@doctrine.orm.entity_manager"
+        container: "@service_container"
